dagster_pipeline/requirements.txt
dagster
sqlalchemy
dagster-webserver
psycopg2-binary
httpx
python-dotenv
pandas